[#18] BottomDetailRootView 분리
